#83 [add] getProfileSet을 ProfileFragment로 기고 HomyType으로 변경
