+++ app/src/main/java/hous/release/android/presentation/profile/ProfileFragment.kt
@@ -9,8 +9,9 @@ import hous.release.android.R
 import hous.release.android.databinding.FragmentProfileBinding
 import hous.release.android.presentation.profile.adapter.ProfilePersonalityAdapter
 import hous.release.android.util.binding.BindingFragment
-import hous.release.android.util.getProfileSet
+import hous.release.domain.entity.HomyType
 import hous.release.domain.entity.PersonalityInfo
+import hous.release.domain.entity.ProfileSet
 
 @AndroidEntryPoint
 class ProfileFragment : BindingFragment<FragmentProfileBinding>(R.layout.fragment_profile) {
@@ -77,6 +78,41 @@ class ProfileFragment : BindingFragment<FragmentProfileBinding>(R.layout.fragmen
                     )
                 }
             }
+        }
+    }
+
+    fun getProfileSet(colorType: String): ProfileSet {
+        return when (HomyType.valueOf(colorType)) {
+            HomyType.RED -> ProfileSet(
+                colorBg = R.color.hous_red_profile,
+                colorText = R.color.hous_red,
+                personality = R.string.personality_red
+            )
+            HomyType.YELLOW -> ProfileSet(
+                colorBg = R.color.hous_yellow,
+                colorText = R.color.hous_yellow,
+                personality = R.string.personality_yellow
+            )
+            HomyType.GREEN -> ProfileSet(
+                colorBg = R.color.hous_green,
+                colorText = R.color.hous_green,
+                personality = R.string.personality_green
+            )
+            HomyType.BLUE -> ProfileSet(
+                colorBg = R.color.hous_blue,
+                colorText = R.color.hous_blue,
+                personality = R.string.personality_blue
+            )
+            HomyType.PURPLE -> ProfileSet(
+                colorBg = R.color.hous_purple,
+                colorText = R.color.hous_purple,
+                personality = R.string.personality_purple
+            )
+            HomyType.GRAY -> ProfileSet(
+                colorBg = -1,
+                colorText = -1,
+                personality = -1
+            )
         }
     }
